Moved `basal_plan_home_page.dart` to a dedicated `pages` folder.
